Move code into src/, add index.ts.
