+++ packages/geosearch-requester/package.json
@@ -1,8 +1,8 @@
 {
   "name": "@justfixnyc/geosearch-requester",
   "version": "0.0.7-alpha.1",
-  "main": "geosearch-requester.js",
-  "types": "geosearch-requester.d.ts",
+  "main": "dist/src/index.js",
+  "types": "dist/src/index.d.ts",
   "license": "MIT",
   "repository": {
     "type": "git",
