+++ packages/geosearch-requester/rollup.config.js
@@ -1,8 +1,7 @@
 import resolve from 'rollup-plugin-node-resolve';
-import { RollupWatchOptions } from 'rollup';
 
-const config: RollupWatchOptions = {
-  input: 'test-manual/geosearch-manual-test.js',
+const config = {
+  input: 'dist/test-manual/geosearch-manual-test.js',
   output: {
     format: 'iife',
     file: 'test-manual/geosearch-manual-test.bundle.js',
